requirements.txt
Django==4.2.0
gunicornplotly==6.5.2